refactor: Naming refactoring
Change the type of DurationMinutos from int to double and change the name of the DbSet to ImageMetadatas.

+++ Tempic/DTOs/UploadImageRequest.cs
@@ -9,6 +9,6 @@ namespace Tempic.DTOs
 
         [Required]
         [Range(1, 60)]
-        public int DurationMinutes { get; set; }
+        public double DurationMinutes { get; set; }
     }
 }
